move index hbs, js, less out of the build directory
