+++ js/index.less
@@ -1,5 +1,5 @@
 @import (inline) 'node_modules/normalize.css/normalize.css';
-@import (inline) '../../node_modules/prismjs/themes/prism.css';
+@import (inline) 'node_modules/prismjs/themes/prism.css';
 
 /*!
  *  Font Awesome 4.7.0 by @davegandy - http://fontawesome.io - @fontawesome
